Check-in

+++ Persistence_Layer/Interfaces/ITransaction.cs
@@ -6,7 +6,7 @@ namespace Persistence_Layer.Interfaces
     public interface ITransaction
     {
         int AccountNo { get; set; }
-        decimal Amount { get; set; }
+        double Amount { get; set; }
         string Description2 { get; set; }
         DateTime TransactionDate { get; set; }
         TransactionType TransactionType { get; set; }
